Refactor maths exprs in QueryPlanSerde
